requirements.txt
fastapi==0.112.0
uvicorn==0.30.6

# PDF Processing
pdfplumber==0.11.3

# OCR & Image
paddleocr==2.8.1
opencv-python-headless==4.10.0.84

# Vector Store & Cosine Calculation
chromadb==0.5.5
numpy==1.26.4
torch==2.4.0

# HTTP
urllib3==2.2.2
colorama==0.4.6

# OpenAI
langchain_openai==0.1.22

# Sentence Transformer
sentence-transformers==3.0.1
transformers==4.44.0

# LLM & RAG
langchain[challenges, prompts]==0.2.14
langchain-huggingface==0.0.3
langchain_community==0.2.12
rank_bm25==0.2.2
bitsandbytes==0.42.0
pytesseract==0.3.10
pandas==2.2.2

# Streamlit
streamlit==1.26.0
requests==2.32.3
pydantic==2.8.2
nltk==3.8.1

# Utility
python-dotenv==1.0.1
tqdm==4.66.5
aiohttp==3.10.3
async-timeout==4.0.3
langchain-core==0.2.33
langchain-text-splitters==0.2.2
langsmith==0.1.99
pyyaml==6.0.2
sqlalchemy==2.0.32
tenacity==8.5.0
python-multipart==0.0.9
tabulate==0.9.0